Move pagination appender to BlockControls
