+++ src/utils/get-icon/index.js
@@ -235,4 +235,8 @@ export default function getIcon( icon ) {
 	if ( 'query-params' === icon ) {
 		return <svg viewBox="0 0 2876 2876" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="2"><path d="M2717.022 158.925c211.903 211.928 211.903 2346.168 0 2558.097-211.928 211.893-2346.169 211.893-2558.097 0-211.9-211.929-211.9-2346.17 0-2558.097 211.928-211.9 2346.169-211.9 2558.097 0ZM1786.4 1975.975c-155.978 116.55-349.419 185.565-558.85 185.565-515.969 0-934.872-418.9-934.872-934.872 0-515.968 418.903-934.868 934.872-934.868 515.969 0 934.872 418.9 934.872 934.868 0 211.963-70.694 407.541-189.725 564.372l508.869 508.87c25.609 25.608 25.609 67.196 0 92.805l-92.81 92.81c-25.61 25.61-67.197 25.61-92.806 0l-509.55-509.55ZM1227.55 529.253c384.916 0 697.419 312.5 697.419 697.415 0 384.916-312.503 697.42-697.419 697.42s-697.419-312.504-697.419-697.42c0-384.915 312.503-697.415 697.419-697.415Zm-28.434 1248.965c-43.153 0-77.044-11.55-101.666-34.65-24.625-23.1-36.934-54.703-36.934-94.812 0-40.616 12.309-72.472 36.934-95.572 24.622-23.1 58.513-34.65 101.666-34.65s76.915 11.55 101.284 34.65c24.369 23.1 36.553 54.956 36.553 95.572 0 40.11-12.184 71.712-36.553 94.812-24.369 23.1-58.131 34.65-101.284 34.65Zm-346.504-811.04c1.016-47.216 9.9-90.497 26.657-129.844 16.753-39.347 40.868-73.11 72.347-101.284 31.475-28.178 69.934-50.138 115.371-65.875 45.438-15.738 97.35-23.607 155.735-23.607 54.325 0 103.697 7.235 148.122 21.704 44.422 14.468 82.5 34.778 114.231 60.921 31.731 26.147 56.225 57.372 73.487 93.672 17.263 36.3 25.894 76.535 25.894 120.703 0 57.37-13.71 106.616-41.125 147.741-27.415 41.122-68.031 77.675-121.847 109.66-25.384 14.724-46.325 28.306-62.825 40.743-16.5 12.438-29.7 25.131-39.6 38.078-9.9 12.944-16.88 26.653-20.943 41.122-4.063 14.469-6.094 30.844-6.094 49.119v40.362h-201.806l-1.525-41.884c-2.538-30.462-1.775-57.878 2.284-82.247 4.063-24.369 11.678-46.581 22.847-66.634 11.169-20.053 26.019-38.46 44.55-55.213s40.997-32.747 67.397-47.978c24.369-14.722 44.803-28.303 61.303-40.74 16.503-12.438 29.703-25.004 39.603-37.697 9.9-12.694 16.878-26.02 20.94-39.982 4.063-13.962 6.094-29.828 6.094-47.596 0-17.77-3.428-34.016-10.28-48.738-6.854-14.722-16.629-27.416-29.32-38.078-12.693-10.66-27.925-19.038-45.693-25.131-17.77-6.09-37.57-9.138-59.4-9.138-44.17 0-79.582 12.438-106.235 37.316-26.656 24.878-41.25 58.384-43.79 100.525H852.612Z" /></svg>;
 	}
+
+	if ( 'add-pagination' === icon ) {
+		return <svg fill="none" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M18 8.2h-5.2V3h-1.6v5.2H6v1.6h5.2V15h1.6V9.8H18V8.2ZM8.5 17.5h-3a1.5 1.5 0 0 0 0 3h3a1.5 1.5 0 0 0 0-3ZM15 19a1.5 1.5 0 1 0-3 0 1.5 1.5 0 0 0 3 0ZM20 19a1.5 1.5 0 1 0-3 0 1.5 1.5 0 0 0 3 0Z" fill="#000" /></svg>;
+	}
 }
